Updated Spell Script, Fixed Spell Transforms, Spell Move Forward
Made the spells correctly move forward and gave them a single
SpellSource object to position newly casted spells.

Aim is now used to cast spells, and spells are automatically readied.
Trying to aim with an unreadied spell clears the command list.

The Spell Interface is gone and is now an abstract object.

Added a Debug Script that draws a line on the object facing forward. We
can mess with this as we wish.

Added *.cs.meta files to the .gitignore