Gitignore, LevelTwo, Committing Assets
Updated Gitignore to have a TestAssets folder so we can throw stuff in there. The plugins folder is so that we don't need to commit files for plugins we install.

Created level two. It's still a test, and needs stuff to happen there. The terrain is sorted. Tweaks to come later

Added a bunch of materials, textures and a few other usables. Not everything is used yet, but could be used in the future. Will see what we need for the situations to come.